Fix meta tags, migrate changelog.mdx
