+++ packages/web/docs/src/content/docs/self-hosting/changelog.mdx
@@ -1,30 +1,22 @@
+---
+title: Self Hosting Changelog
+description: Changelog for self-hosting Hive
+---
+
 import * as fs from 'node:fs/promises'
 import * as path from 'node:path'
-import { buildDynamicMDX } from 'nextra/remote'
+import { MDXRemote } from 'nextra/mdx-remote'
 import { RemoteContent } from '@theguild/components'
+import { compileMdx, useMDXComponents } from '@theguild/components/server'
 
-export async function getStaticProps() {
+export async function compileChangelogMdx() {
   // NOTE: if you add any white space between lines in this function, prettier will destroy this file
-  // Unfortunately, it seems impossible to use `<RemoteContent />` within a default export of a `.tsx` file.
-  const data = (
-    await fs.readFile(path.join(process.cwd(), '../../../deployment/CHANGELOG.md'), 'utf-8')
+  let source = await fs.readFile(
+    path.join(process.cwd(), '../../../deployment/CHANGELOG.md'),
+    'utf-8'
   )
-    // remove the heading
-    .split('\n')
-    .slice(1)
-    .join('\n')
-  const dynamicMdx = await buildDynamicMDX(data)
-  return {
-    props: {
-      ...dynamicMdx,
-      __nextra_dynamic_opts: {
-        title: 'Self Hosting Changelog',
-        frontMatter: {
-          description: 'Changelog for self-hosting Hive'
-        }
-      }
-    }
-  }
+  source = source.split('\n').slice(1).join('\n') // remove the heading
+  return compileMdx(source)
 }
 
 # Self-Hosting Changelog
@@ -33,4 +25,4 @@ Here you can find all the recent improvements and updates about the self-hosting
 images published to the GitHub registry using the same tags. For a high-level overview of
 improvements please refer to [our product updates](/product-updates).
 
-<RemoteContent />
+<MDXRemote compiledSource={await compileChangelogMdx()} />
